#2396 Move to grunt/api folder and separate entry/grunt modules
