--- grunt/tasks/setenv.js
@@ -1,5 +1,0 @@
-module.exports = function(grunt) {
-    grunt.registerTask('setenv', function(env) {
-        grunt.config('env', env);
-    });
-};
