+++ grunt/tasks/setoutput.js
@@ -1,0 +1,5 @@
+module.exports = function(grunt) {
+    grunt.registerTask('setoutput', function(outputConfig) {
+        grunt.config('outputConfig', outputConfig);
+    });
+};
